Continuing implementation of Item classes
