+++ include/constants.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace constants
+{
+	constexpr int undefined = -1;
+	constexpr char max_life_points = 20;
+	constexpr char max_lamp_points = 60;
+	constexpr char max_owned_items = 10;
+}
