--- include/constants/r136.h
@@ -1,8 +1,0 @@
-#pragma once
-
-// constants/r136.h
-
-constexpr int undefined = -1;
-constexpr char max_life_points = 20;
-constexpr char max_lamp_points = 60;
-constexpr char max_owned_items = 10;
