+++ include/gamedata.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "types/base.h"
+
+namespace game_data
+{
+	bool load(CoreData& core);
+	bool save(CoreData& core);
+	void remove();
+}
