+++ include/general.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "types/general.h"
+#include "inlines/general.h"
