--- include/init.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "types/r136.h"
-
-bool set_room_connections(std::vector<Room*> rooms);
