+++ include/random.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "types/random.h"
+#include "inlines/random.h"
