+++ include/startup.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "types/base.h"
+
+namespace startup
+{
+	void run_intro();
+	void show_splashscreen(void);
+	void show_start_message(void);
+	void initialize(CoreData& core);
+}
