+++ include/types/animates.h
@@ -2,7 +2,7 @@
 
 // types/animate.h
 
-#include "types/r136.h"
+#include "types/base.h"
 
 class Barbecue : public Animate
 {
