+++ include/types/general.h
@@ -1,0 +1,9 @@
+#pragma once
+
+// types/general.h
+
+namespace general
+{
+	void force_exit(void);
+	void sleep_ms(int count);
+}
