+++ src/items/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(
+	${PROJECT_NAME} 
+	PRIVATE 
+	bandage.cpp
+	flashlight.cpp
+	tnt.cpp
+)
